+++ src/extension.ts
@@ -7,11 +7,14 @@ import { sep } from 'path';
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 
+import * as clipboard from 'copy-paste';
+
 import PackReader from './util/PackReader';
 import PathConstants from './constant/PathConstants';
 import CopyTask from './model/CopyTask';
 import Builder from './util/Builder';
 import FsHelper from './util/FsHelper';
+import { rejects } from 'assert';
 
 const packReader = new PackReader();
 
@@ -74,30 +77,87 @@ export function activate(context: vscode.ExtensionContext) {
 		buildFunction(packReader.projects, packReader.tasks);
 	});
 
-	let watch = vscode.commands.registerCommand('node-workspace-builder.watchProject', (e) => {
-		if (/node_modules/.test(e.fsPath)) {
-			vscode.window.showWarningMessage('This is a dependency installation folder. Workspace builder will not watch this.');
-			return;
-		}
-		const isPackJson = e.fsPath.includes(PathConstants.PACK_JSON);
-		const includesPackJson = FsHelper.exists(`${e.fsPath}${sep}${PathConstants.PACK_JSON}`);
-		if (!isPackJson && !includesPackJson) {
-			vscode.window.showWarningMessage('There is no package.json file found. This folder is not a node project folder.');
-			return;
-		}
+	let watch = vscode.commands.registerCommand('node-workspace-builder.watchProject', (uri: vscode.Uri) => {
+		new Promise<Array<vscode.Uri>>((resolve, reject) => {
+			if (uri === undefined) {
+				vscode.commands.executeCommand('copyFilePath').then(value => {
+					const paths = clipboard.paste().split('\n');
+					const uris = paths.map(m => vscode.Uri.file(m)).filter(f => FsHelper.exists(f.fsPath) && !f.fsPath.includes('vs_code_welcome_page'));
+					uris && uris.length ? resolve(uris) : reject(new Error('No file or folder selected while running command directly.'));
+				});
+			} else {
+				resolve([uri]);
+			}
+		}).then((realUris: Array<vscode.Uri>) => {
+			const promises = realUris.filter(f => {
+				if (/node_modules/.test(f.fsPath)) {
+					output.appendLine('This is a dependency installation folder. Workspace builder will not watch this: ' + f.fsPath);
+				}
+				const isPackJson = f.fsPath.includes(PathConstants.PACK_JSON);
+				const includesPackJson = FsHelper.exists(`${f.fsPath}${sep}${PathConstants.PACK_JSON}`);
+				if (!isPackJson && !includesPackJson) {
+					output.appendLine('There is no package.json file found. This folder is not a node project folder: ' + f.fsPath);
+				}
+				return !/node_modules/.test(f.fsPath) && (isPackJson || includesPackJson);
+			}).map(realUri => {
+				const isPackJson = realUri.fsPath.includes(PathConstants.PACK_JSON);
+				const file = isPackJson
+					? realUri.fsPath.replace(PathConstants.PACK_JSON, PathConstants.PLACEHOLDER)
+					: `${realUri.fsPath}${sep}${PathConstants.PLACEHOLDER}`;
 
-		const file = isPackJson
-			? e.fsPath.replace(PathConstants.PACK_JSON, PathConstants.PLACEHOLDER)
-			: `${e.fsPath}${sep}${PathConstants.PLACEHOLDER}`;
-
-		FsHelper.writeFile(file, '');
-		prepare().then(() => {
+				FsHelper.writeFile(file, '');
+				return Promise.resolve();
+			});
+			return Promise.all(promises).then(prepare);
+		}).then(() => {
 			buildFunction(packReader.projects, packReader.tasks);
+		}).catch(err => {
+			vscode.window.showWarningMessage(err.message);
+		});
+	});
+
+	let buildProject = vscode.commands.registerCommand('node-workspace-builder.buildProject', (uri) => {
+		new Promise<Array<vscode.Uri>>((resolve, reject) => {
+			if (uri === undefined) {
+				vscode.commands.executeCommand('copyFilePath').then(value => {
+					const paths = clipboard.paste().split('\n');
+					const uris = paths.map(m => vscode.Uri.file(m)).filter(f => FsHelper.exists(f.fsPath) && !f.fsPath.includes('vs_code_welcome_page'));
+					uris && uris.length ? resolve(uris) : reject(new Error('No file or folder selected while running command directly.'));
+				});
+			} else {
+				resolve([uri]);
+			}
+		}).then((realUris: Array<vscode.Uri>) => {
+			const promises = realUris.filter(f => {
+				if (/node_modules/.test(f.fsPath)) {
+					output.appendLine('This is a dependency installation folder. Workspace builder will not watch this: ' + f.fsPath);
+				}
+				const isPackJson = f.fsPath.includes(PathConstants.PACK_JSON);
+				const includesPackJson = FsHelper.exists(`${f.fsPath}${sep}${PathConstants.PACK_JSON}`);
+				if (!isPackJson && !includesPackJson) {
+					output.appendLine('There is no package.json file found. This folder is not a node project folder: ' + f.fsPath);
+				}
+				return !/node_modules/.test(f.fsPath) && (isPackJson || includesPackJson);
+			}).map(realUri => {
+				const isPackJson = realUri.fsPath.includes(PathConstants.PACK_JSON);
+				const file = isPackJson ? realUri.fsPath.replace(PathConstants.PACK_JSON, '') : `${realUri.fsPath}${sep}`;
+				return Promise.resolve(file);
+			});
+			return Promise.all(promises);
+		}).then((files: Array<string>) => {
+			const tasks = new Array<CopyTask>();
+			files.forEach(file => {
+				tasks.splice(tasks.length, 0, ...packReader.tasks.filter(f => f.projectDepPath.includes(file)));
+			});
+			buildFunction(files, tasks);
+		}).catch(err => {
+			vscode.window.showErrorMessage(err.message);
 		});
 	});
 
 	context.subscriptions.push(build);
 	context.subscriptions.push(watch);
+	context.subscriptions.push(buildProject);
 
 	vscode.workspace.onDidSaveTextDocument((e: vscode.TextDocument) => {
 		const configuration = vscode.workspace.getConfiguration('node-workspace-builder');